% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPRS.R
\name{Binary_agg.fun}
\alias{Binary_agg.fun}
\title{binomial aggregation rule}
\usage{
Binary_agg.fun(
  B,
  Xtest.pfile,
  y.test,
  nsplit = 4,
  const = 2,
  phenotype,
  covariates = NULL
)
}
\arguments{
\item{B}{a matrix for column-binded possible beta}

\item{Xtest.pfile}{genofile for test. The PLINK 2.0 pgen file that contains genotype.
We assume the existence of Xtest.pfile.{pgen,pvar.zst,psam}.}

\item{y.test}{phenofile for test}

\item{nsplit}{split number of X and beta that we can product part by part to reduce memory use. default is 4}

\item{const}{}

\item{phenotype}{the name of the phenotype. Must be the same as the corresponding column name in
the phenotype file.}

\item{covariates}{a character vector containing the names of the covariates included in the lasso
fitting, whose coefficients will not be penalized. The names must exist in the
column names of the phenotype file.}
}
\description{
binomial aggregation rule
}
