% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPRS.R
\name{bind_phe}
\alias{bind_phe}
\title{bind phe}
\usage{
bind_phe(nameLis, out)
}
\arguments{
\item{nameLis}{nameLis of phenofile. We assume the existence of nameLis.{phe}}

\item{out}{output phenofile name}
}
\description{
bind phe
}
