% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPRS.R
\name{bindVCF_py}
\alias{bindVCF_py}
\title{bind vcf using python}
\usage{
bindVCF_py(nameLis, out)
}
\arguments{
\item{nameLis}{We assume the existence of nameLis.{vcf}.}

\item{out}{output vcf file name}
}
\description{
bind vcf using python
}
