% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPRS.R
\name{keep_pfile}
\alias{keep_pfile}
\title{select given sample from pgen file and generate new pgen files}
\usage{
keep_pfile(configs, genotype.pfile, keepnamefile, out.pfile)
}
\arguments{
\item{configs}{a list of other config parameters.}

\item{genotype.pfile}{the PLINK 2.0 pgen file that contains genotype.
We assume the existence of genotype.pfile.{pgen,pvar.zst,psam}.}

\item{keepnamefile}{a file with FID and IID of samples to keep names.}

\item{out.pfile}{the output PLINK 2.0 pgen file name that contains genotype.}
}
\description{
select given sample from pgen file and generate new pgen files
}
