% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPRS.R
\name{mse.fun}
\alias{mse.fun}
\title{calculate mse}
\usage{
mse.fun(beta, est, X.test = NULL)
}
\arguments{
\item{beta}{}

\item{est}{estimated beta}

\item{X.test}{X(geno) matrix that can be added to calculate pred.err}
}
\description{
calculate mse
}
