% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPRS.R
\name{pgenGen}
\alias{pgenGen}
\title{transformation of vcf file to pgen file}
\usage{
pgenGen(out_pgen_name, vcf.file, id_type)
}
\arguments{
\item{out_pgen_name}{the output PLINK 2.0 pgen file named as out_pgen_name.{pgen,pvar.zst,psam}.}

\item{vcf.file}{input VCF files}

\item{id_type}{"underline" or "double", depend on the sample names in VCF files
if the sample name is 'FID_IID', choose id_type="underline"
if the sample name is 'IID' alone , choose id_type="double"}
}
\description{
transformation of vcf file to pgen file
}
