% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPRS.R
\name{pheGen}
\alias{pheGen}
\title{transformation of phenofile with FID and IID to snpnet phenofile with split.col}
\usage{
pheGen(
  ori_phefile,
  phenotype = "QPHE",
  covariates = NULL,
  offset = NULL,
  split.col = "split",
  out_phefile,
  split_vec = c(0.9, 0.1)
)
}
\arguments{
\item{ori_phefile}{original phenofile name. With FID, IID and phenotypes in the file, with no header}

\item{phenotype}{defile output column names of phenotype columns. "QPHE" default}

\item{covariates}{a character vector containing the names of the covariates included in the lasso
fitting, whose coefficients will not be penalized. The names must exist in the
column names of the phenotype file.}

\item{offset}{a character vector containing the names of the offset included in the lasso
fitting. The names must exist in the column names of the phenotype file.}

\item{split.col}{defile output column names of split columns. "split" default}

\item{out_phefile}{output phenofile name}

\item{split_vec}{proportions to split the data into training set and validating set
to find the optimal lambda
c(0.9, 0.1) for "train" and "val" default}
}
\description{
transformation of phenofile with FID and IID to snpnet phenofile with split.col
}
