% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPRS.R
\name{read_geno}
\alias{read_geno}
\title{read a pfile and save it as a matrix}
\usage{
read_geno(new_genotype_file, per_idx = NULL, select_SNP = NULL)
}
\arguments{
\item{new_genotype_file}{the PLINK 2.0 pgen file that contains genotype.
We assume the existence of genotype.pfile.{pgen,pvar.zst,psam}.}

\item{per_idx}{index of selected persons}

\item{select_SNP}{index of selected SNPs}
}
\description{
read a pfile and save it as a matrix
}
