% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPRS.R
\name{remove_anno}
\alias{remove_anno}
\title{remove all the annotations that start with '##'}
\usage{
remove_anno(filename)
}
\arguments{
\item{filename}{}
}
\description{
remove all the annotations that start with '##'
}
