% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPRS.R
\name{rsq_cal}
\alias{rsq_cal}
\title{return r^2 of prediction (under linear assumption)}
\usage{
rsq_cal(est, X.test, y_real)
}
\arguments{
\item{est}{estimated beta}

\item{X.test}{X(geno) matrix}

\item{y_real}{real y (phenotypes)}
}
\description{
return r^2 of prediction (under linear assumption)
}
