% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPRS.R
\name{split_vcf}
\alias{split_vcf}
\title{split the files as training set and testing set, I.til can be non-consecutive integers}
\usage{
split_vcf(
  vcf.file = vcf.file,
  genotype.pfile = genotype.pfile,
  phenotype.file = phenotype.file,
  configs = configs,
  get_path = get_path,
  I.til = I.til,
  n = sum(n.vec),
  outname = c("test", "train")
)
}
\arguments{
\item{vcf.file}{geno file in vcf format}

\item{genotype.pfile}{the PLINK 2.0 pgen file that contains genotype.
We assume the existence of genotype.pfile.{pgen,pvar.zst,psam}.}

\item{phenotype.file}{the path of the file that contains the phenotype values and can be read as
as a table. There should be FID (family ID) and IID (individual ID) columns
containing the identifier for each individual, and the phenotype column(s).
(optional) some covariate/offset columns and a column specifying the
training/validation split can be included in this file.}

\item{configs}{a list of other config parameters.}

\item{get_path}{dir for saving the intermediate files and result files}

\item{I.til}{A vector for selected samples to be a testing set. Testing set named as outname\link{1}}

\item{n}{total sample size}

\item{outname}{output filename: ${outname}.pfile, ${outname}.phe
c("test", "train") is default}
}
\description{
split the files as training set and testing set, I.til can be non-consecutive integers
}
