% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPRS.R
\name{ST.init}
\alias{ST.init}
\title{single-task initialization}
\usage{
ST.init(
  genotype.pfile = genotype.pfile,
  phenotype.file = phenotype.file,
  phenotype = phenotype,
  covariates = NULL,
  offset = NULL,
  split.col = NULL,
  configs = configs,
  nlambda = 30,
  lambda.min.ratio = 0.01,
  p = NULL,
  family,
  alpha = 1,
  shortcut = T
)
}
\arguments{
\item{genotype.pfile}{the PLINK 2.0 pgen file that contains genotype.
We assume the existence of genotype.pfile.{pgen,pvar.zst,psam}.}

\item{phenotype.file}{the path of the file that contains the phenotype values and can be read as
as a table. There should be FID (family ID) and IID (individual ID) columns
containing the identifier for each individual, and the phenotype column(s).
(optional) some covariate/offset columns and a column specifying the
training/validation split can be included in this file.}

\item{phenotype}{the name of the phenotype. Must be the same as the corresponding column name in
the phenotype file.}

\item{covariates}{a character vector containing the names of the covariates included in the lasso
fitting, whose coefficients will not be penalized. The names must exist in the
column names of the phenotype file.}

\item{offset}{a character vector containing the names of the offset included in the lasso
fitting. The names must exist in the column names of the phenotype file.}

\item{split.col}{the column name in the phenotype file that specifies the membership of individuals to
the training or the validation set. The individuals marked as "train" and "val" will
be treated as the training and validation set, respectively. When specified, the
model performance is evaluated on both the training and the validation sets.}

\item{configs}{a list of other config parameters.}

\item{nlambda}{the number of lambda values - default is 50}

\item{lambda.min.ratio}{smallest value for lambda, as a fraction of lambda.max, the (data derived) entry value,
i.e. the smallest value for which all coefficients are zero. The default
depends on the sample size nobs relative to the number of actual variables
nvars (after QC filtering). The default is 0.01. A very small value of lambda.min.ratio
will lead to a saturated fit in the nobs < nvars case.}

\item{p}{number of SNPs in genofile}

\item{family}{the type of the phenotype: "gaussian", "binomial".}

\item{alpha}{the elastic-net mixing parameter, where the penalty is defined as
alpha * ||beta||_1 + (1-alpha)/2 * ||beta||_2^2. alpha = 1 corresponds to the lasso penalty,
while alpha = 0 corresponds to the ridge penalty.}

\item{shortcut}{whether to avoid using tuned lambda to caltulate the parameters on all samples again. If the
sample size is large enough, let shortcut = T can save computational costs. #'}
}
\description{
single-task initialization
}
