% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPRS.R
\name{Xbeta_product}
\alias{Xbeta_product}
\title{Xt\%*\%beta}
\usage{
Xbeta_product(new_genotype_file, beta, p, nsplit = NULL)
}
\arguments{
\item{new_genotype_file}{the PLINK 2.0 pgen file that contains genotype.
We assume the existence of genotype.pfile.{pgen,pvar.zst,psam}.}

\item{beta}{beta value}

\item{p}{the number of SNPs in pgen file}

\item{nsplit}{split number of X and beta that we can product part by part to reduce memory use}
}
\description{
Xt\%*\%beta
}
