% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransPRS.R
\name{Xy_product}
\alias{Xy_product}
\title{only calculate Xt\%*\%y of appointed samples}
\usage{
Xy_product(new_genotype_file, y, select_per, phenotype)
}
\arguments{
\item{new_genotype_file}{the PLINK 2.0 pgen file that contains genotype.
We assume the existence of genotype.pfile.{pgen,pvar.zst,psam}.}

\item{y}{phenotypes}

\item{select_per}{index of person selected}

\item{phenotype}{the colnames of phenotypes}
}
\description{
only calculate Xt\%*\%y of appointed samples
}
